move contract_fixture to tests/core/vm
