+++ tests/core/vm/contract_fixture.py
